Subring map (#5)
Maps some ring R with a large degree `d*k` (e.g. 512 of Falcon) into a
set of `k` rings S of degree `d`. This set is represented through the
struct `SplitRing`. Each subring S is a `SuitableRing` usable in
LatticeFold.
The map follows the bijection provided in
[LNPS21](https://eprint.iacr.org/2021/1575) Section 2.8.
An R1CS example of a multiplication between two `SplitRing`s is
provided. This will be then refactored in another PR for easier
manipulation/R1CS construction.