R1CS norm bound example (#6)
Adds R1CS which verifies that some poly has l2 norm lower than a power
of 2 bound.
Given that direct ops like `<`, `>=`, are not available in R1CS, the
method employed here is through bit decomposition, where the computed l2
norm is shown to be representable using up to `log2(bound)` bits.
No roots are computed here, so we only operate on the squared norm /
bound.